Organize note commitment tests and test vectors
