+++ zebra-chain/src/sapling/tests.rs
@@ -1,2 +1,4 @@
 mod preallocate;
 mod prop;
+mod test_vectors;
+mod vectors;
